GEM-2341: Cherry-picked 947c4db09d6463d810d5c6799ec0b22451ea6e1f GEODE-6215: Making session meta-data region distributed (#3012)
The meta data region used for session replication was converted to
Scope.LOCAL by mistake in 7ca571c5. This broke the logic to
automatically create the session region on all peers when starting up
the session replication service.

Changing the scope back to DISTRIBUTED_ACK for this metadata region, and
adding tests that session region creation works.

* Rolling upgrade test for session modules
* Disabling rolling upgrade test on java 11. The
  Tomcat8ClientServerRollingUpgradeTest won't work with geode versions less
  and 1.8 on java 11.

Co-Authored-By: Dan Smith <dsmith@pivotal.io>